requirements.txt
alembic==1.13.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
outcome==1.3.0.post0
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.1
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.18.2
